feat: further dynamic content changes
